lambdas/load_staging/requirements.txt
boto3==1.35.72
google-cloud-bigquery==3.29.0
pandas==2.2.2
pyarrow==19.0.0